LPD-31881 Alow our fallback error page to handle the RuntimeException
Based on the Servlet specification (SRV.8.5), a servlet that is the
target of a request dispatcher may throw a runtime exception, which
should then be propagated to the calling servlet.

Since this portion of code provides custom error handling for the
calling servlet: Main Servlet, we will catch the RuntimeException
and account for the exception when serving our fallback error page.

Note: Without this commit, test case: testCustomErrorPageFallback
would fail if the following portal property is set to true:
direct.servlet.context.enabled (enabled by default).